Move github.com/sigstore/protobuf-specs users into a separate subpackage
Nothing else in this repository currently uses sigstore/protobuf-specs,
notably prodution users of pkg/client don't use it. So, move the new TLE
utilities into a separate subpackage, to make clients smaller.

Only moves unchanged code, should not change behavior.
