+++ cmd/rekor-cli/app/format/wrap.go
@@ -21,8 +21,8 @@ import (
 
 	rekor_pb "github.com/sigstore/protobuf-specs/gen/pb-go/rekor/v1"
 
-	"github.com/sigstore/rekor/pkg/client"
 	"github.com/sigstore/rekor/pkg/log"
+	tleutils "github.com/sigstore/rekor/pkg/tle"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -51,7 +51,7 @@ func WrapCmd(f formatCmd) CobraCmd {
 			fmt.Println(toJSON(obj))
 		case "tle":
 			if tle, ok := obj.(*rekor_pb.TransparencyLogEntry); ok {
-				json, err := client.MarshalTLEToJSON(tle)
+				json, err := tleutils.MarshalTLEToJSON(tle)
 				if err != nil {
 					log.CliLogger.Fatalf("error converting to transparency log entry: %v", err)
 				}
